Move lots of stuff to _action.py and introduce explicit filters, parameters, filter_interaction arguments instead of **inputs
