+++ vizro-core/src/vizro/actions/_on_page_load_action.py
@@ -1,26 +1,24 @@
 """Pre-defined action function "_on_page_load" to be reused in `action` parameter of VizroBaseModels."""
 
-from pydantic import field_validator
-from typing import Any, Optional, Callable
+from typing import Any
 
-from dash import ctx, Output
+from dash import State, ctx
 
 from vizro.actions._actions_utils import _get_modified_page_figures
-
-
-from vizro.managers._model_manager import ModelID, model_manager
+from vizro.managers._model_manager import ModelID
 from vizro.models._action._action import NewAction
-
-from vizro.models.types import capture
-
-from vizro.models import Action
 
 
 # TODO NOW: rename apply_controls or similar. Docstring. Tidy comments in arguments.
 class _on_page_load(NewAction):
     targets: list[ModelID]
 
-    def __call__(self, **inputs: dict[str, Any]) -> dict[ModelID, Any]:
+    def __call__(
+        self,
+        filters: list[State],
+        parameters: list[State],
+        filter_interaction: list[dict[str, State]],
+    ) -> dict[ModelID, Any]:
         """Applies controls to charts on page once the page is opened (or refreshed).
 
         Args:
@@ -32,7 +30,6 @@ class _on_page_load(NewAction):
         Returns:
             Dict mapping target component ids to modified charts/components e.g. {'my_scatter': Figure({})}
         """
-
         # TODO NOW: comment about how filters, paramters, filter_interaction currently ignored but will be used in
         #  future when do pattern matching callback.
